Restore search to EPM list page
